+++ packages/jazz-tools/src/implementation/createContext.ts
@@ -11,10 +11,16 @@ import {
 } from "cojson";
 import { Account, AccountClass, ID } from "../internal.js";
 
+export type Credentials = {
+    accountID: ID<Account>;
+    secret: AgentSecret;
+};
+
 export type AuthResult =
     | {
           type: "existing";
-          credentials: { accountID: ID<Account>; secret: AgentSecret };
+          credentials: Credentials;
+          saveCredentials?: (credentials: Credentials) => Promise<void>;
           onSuccess: () => void;
           onError: (error: string | Error) => void;
           logOut: () => void;
@@ -23,10 +29,7 @@ export type AuthResult =
           type: "new";
           creationProps: { name: string };
           initialSecret?: AgentSecret;
-          saveCredentials: (credentials: {
-              accountID: ID<Account>;
-              secret: AgentSecret;
-          }) => Promise<void>;
+          saveCredentials: (credentials: Credentials) => Promise<void>;
           onSuccess: () => void;
           onError: (error: string | Error) => void;
           logOut: () => void;
@@ -44,6 +47,7 @@ export const fixedCredentialsAuth = (credentials: {
         start: async () => ({
             type: "existing",
             credentials,
+            saveCredentials: async () => {},
             onSuccess: () => {},
             onError: () => {},
             logOut: () => {},
@@ -168,6 +172,15 @@ export async function createJazzContext<Acc extends Account>(
                     });
 
                     const account = AccountSchema.fromNode(node);
+
+                    if (authResult.saveCredentials) {
+                        await authResult.saveCredentials({
+                            accountID: node.account
+                                .id as unknown as ID<Account>,
+                            secret: node.account.agentSecret,
+                        });
+                    }
+
                     authResult.onSuccess();
 
                     return {