+++ examples/password-manager/CHANGELOG.md
@@ -1,5 +1,29 @@
 # jazz-password-manager
 
+## 0.0.13
+
+### Patch Changes
+
+-   Updated dependencies [cce679b]
+    -   jazz-tools@0.8.15
+    -   jazz-react@0.8.15
+
+## 0.0.12
+
+### Patch Changes
+
+-   Updated dependencies [36273b3]
+    -   jazz-tools@0.8.14
+    -   jazz-react@0.8.14
+
+## 0.0.11
+
+### Patch Changes
+
+-   Updated dependencies [fd011d7]
+    -   jazz-tools@0.8.13
+    -   jazz-react@0.8.13
+
 ## 0.0.10
 
 ### Patch Changes