+++ packages/jazz-react-native-media-images/CHANGELOG.md
@@ -1,5 +1,26 @@
 # jazz-browser-media-images
 
+## 0.8.11
+
+### Patch Changes
+
+-   Updated dependencies [cce679b]
+    -   jazz-tools@0.8.15
+
+## 0.8.10
+
+### Patch Changes
+
+-   Updated dependencies [36273b3]
+    -   jazz-tools@0.8.14
+
+## 0.8.9
+
+### Patch Changes
+
+-   Updated dependencies [fd011d7]
+    -   jazz-tools@0.8.13
+
 ## 0.8.8
 
 ### Patch Changes