+++ packages/jazz-tools/CHANGELOG.md
@@ -1,5 +1,23 @@
 # jazz-tools
 
+## 0.8.15
+
+### Patch Changes
+
+-   cce679b: Export `type Credentials` and mark `AuthResult.saveCredentials` as optional and run if available
+
+## 0.8.14
+
+### Patch Changes
+
+-   36273b3: Cache other usages of Account.fromRaw
+
+## 0.8.13
+
+### Patch Changes
+
+-   fd011d7: Add a cache layer on the loadedAs account reads
+
 ## 0.8.12
 
 ### Patch Changes